+++ ui/screens/hw.screen.tsx
@@ -1,0 +1,9 @@
+import { Box } from 'native-base';
+
+export const HelloWorld = () => {
+	return (
+		<Box>
+			Hello World
+		</Box>
+	);
+}
